Added handler for single activity
New activity handler will validate paths and return an error if invalid.

Had to duplicate the activity template. Need to figure out how
to use a one template for single activities.

Installed slices package. Was able to use the IndexFunc to look up the
index of a struct in a slice. Similar to Array.findIndex() in js.
Will figure out if possible to return the actual struct.